Rename package smallrye.health to microprofile.smallrye.health
